src/app/pyproject.toml
[project]
name = "db-residential-copilot"
dynamic = ["version"]
description = "Databricks Residential Investment Copilot"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.119.0",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.37.0",
    "databricks-sdk>=0.81.0",
    "sqlmodel>=0.0.27",
    "psycopg[binary,pool]>=3.2.11",
    "langchain-core>=0.3.0",
    "langchain-databricks>=0.1.0",
    "langgraph>=0.2.0",
]

[dependency-groups]
dev = [
    "ty>=0.0.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/db_residential_copilot/_version.py"

[tool.hatch.build]
artifacts = ["src/db_residential_copilot/__dist__", "src/db_residential_copilot/_version.py"]
